Refactor: Organize Theme Files and Folders for Material UI Sign-In and Sign-Up Templates
- Restructured the theme files and folders to improve organization and maintainability

- Ensured consistency in styling and theming across both components

- Removed unnecessary code and optimized imports
