+++ frontend/src/pages/sign-in/SignIn.jsx
@@ -127,7 +127,7 @@ export default function SignIn(props) {
             variant="h4"
             sx={{ width: "100%", fontSize: "clamp(2rem, 10vw, 2.15rem)" }}
           >
-            Sign in
+            Sign In
           </Typography>
           <Box
             component="form"
