--- frontend/src/pages/sign-up/theme/customizations/index.jsx
@@ -1,5 +1,0 @@
-export { inputsCustomizations } from "./inputs";
-export { dataDisplayCustomizations } from "./dataDisplay";
-export { feedbackCustomizations } from "./feedback";
-export { navigationCustomizations } from "./navigation";
-export { surfacesCustomizations } from "./surfaces";
