--- frontend/src/pages/sign-up/theme/getSignUpTheme.jsx
@@ -1,21 +1,0 @@
-import { getDesignTokens } from "./themePrimitives";
-import {
-  inputsCustomizations,
-  dataDisplayCustomizations,
-  feedbackCustomizations,
-  navigationCustomizations,
-  surfacesCustomizations,
-} from "./customizations";
-
-export default function getSignUpTheme(mode) {
-  return {
-    ...getDesignTokens(mode),
-    components: {
-      ...inputsCustomizations,
-      ...dataDisplayCustomizations,
-      ...feedbackCustomizations,
-      ...navigationCustomizations,
-      ...surfacesCustomizations,
-    },
-  };
-}
